refactor(utils): move model pricing table out of the CLI layer
`pricing.ts` and its `model-normalizer.ts` dependency lived under
`src/cli/commands/analytics/`, so any non-CLI consumer had to import upward,
against the documented CLI -> Registry -> Plugin -> Core -> Utils direction.
Both are layer-agnostic — a memoized JSON price lookup and a pure string
helper — so they belong in Utils, which every layer may depend on.

Analytics consumers now import them through the `@/utils/` alias, and the
build asset rule copies `src/utils/pricing.json` to `dist/utils/pricing.json`.
No logic changed inside the moved files; only their own import paths.
